Refactored main decode loop to remove check against bytes_left
This is part of the upcoming fast decode commit. The fast path didn't compare
against bytes_left in the main loop, which caused a number of tests to fail
as unexpected end-of-stream errors were raised on the stream.

This change hoists the check for bytes_left from pb_decode_varint32_eof,
up into pb_decode_tag, such that eof is set before an error is raised
when called with bytes_left == 0. This simplifies the main loop from -

while (stream->bytes_left)
    if (!pb_decode_tag())
        ...

to -

while (pb_decode_tag())
    ...

This is also important because bytes_left isn't guaranteed to be valid for callback
streams. This created an edge case where callback streams would set an "io error"
when eof was hit, regardless of if decoding was successful or not. This error could
then mask actual errors generated after the main loop (e.g. missing required fields).

A workaround and accompanying regression test was added for this, and is being
further tracked in issue #1017.